Use data scale for u and v passed to barbs
I was using a screen scaling for u and v, this might be useful for
quivers, where u and v are used to draw, but for wind barbs u and v are
just used to decide what the glyph looks like.

Perhaps we should add in the scale of the barb as an additional argument
as it is for Circle.