Only update time series when visible
Listen for bokeh.event.Tap state and update the store with "position"
data.

SeriesView and MarkDraw listen for changes in state["position"], and
then update the time series and marker location accordingly.

SeriesView can now also listen for changes "time_series_visible", to
prevent unneccesary replotting when the series is invisible.

This commit has some issues. Forest seems to behave strangely when
initially opening. The time seriese will disappear on the first click,
and require two toggles of "time_series_visible" to reappear. I think
this is to do with how I have managed the defaults.